+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.1] - 2021-02-25
+
+### Added
+- feeder_MAXMIND_GeoIP.py [login to https://www.maxmind.com/ and get FREE API License Key]
+
+### Changed
+- GITHUB - fix line 47 / ignore new section: ssh_keys
 
 ## [0.6.0] - 2020-01-22